CultureApp420/project-2025-group/src/requirements.txt
ansible==11.11.0
asgiref==3.9.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
fonttools==4.56.0
kiwisolver==1.4.8
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.2.3
reportlab>=3.6.13
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==6.1.1
Pygments==2.19.1
pyparsing==3.2.1
pyqtgraph==0.13.7
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qrcode>=7.4.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2

gunicorn==23.0.0
dotenv~=0.9.9
python-dotenv~=1.1.1
django-cas-client==1.5.3
django-auth-adfs>=1.15.1

djangorestframework-api-key==3.*
